feat(controls): implement buttons and joysticks, design to parity
Mostly implementation, but I also reorganized a few things. I have been sort of
struggling with control statefulness in React. One of React's theses is
"immutable everything", but the control state 'tree' is inherently mutable.
While I could swap that around to be Redux-based or something along those
lines, a) I didn't want to tightly bind it to a certain framework and b) for
various reasons immutability is not an idiom of most software and systems, and
I don't want to add further obstacles to custom control productivity. Same
reason I'm not using rxjs, in spite of how much easier it would make life!

Anyhow. I introduced a `ResourceHolder` component that "fakes" immutability
be passing in the mutable control's properties as React props, whereas before
they were in the component's very mutable state. I also removed the aliasing
that were in components before, those are just going to lead to confusion
("why can I look up the button's disabled state but not my custom prop?").
This seems to work decently well and hides most complexity from the controls.

Some screenshots:

[https://peet.io/i/17-08-iby1497v3zjarsd4tqcfxlnfts5lcc.gif](![](https://peet.io/i/17-08-iby1497v3zjarsd4tqcfxlnfts5lcc.gif))
[https://peet.io/i/17-08-epbtud9hz0vjxqpl7tiug5hir5g5hl.png](![](https://peet.io/i/17-08-epbtud9hz0vjxqpl7tiug5hir5g5hl.png))
[https://peet.io/i/17-08-x8eqx5wzm4isk9fsebidwzvbnuwlgt.gif](![](https://peet.io/i/17-08-x8eqx5wzm4isk9fsebidwzvbnuwlgt.gif))